Create intl_ee.arb
It's the Estonian translation. Translated by a human (not machine). Some of the word order is changed to make it easier to understand for the reader (because the Estonian language differs from the English language structure). Terms from the Estonian IT-glossary are used (the same terms that are used by official Estonian translations by IT-companies like Google etc).